+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 ## [2.2.2]
 ### Added
 - Compatibility with foundry v10.
+### Changed
+- Now measured templates dynamically target tokens even before placing them.
+- Auto targeting by measured templates is now enabled by default.
 ### Notes
 - Worlds created prior to version v10 must be migrated to v10. This process is irreversible, so as always doing a backup is strongly advised.